requirements.txt
asgiref==3.3.4
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
decorator==4.4.2
Django==3.2.4
google-api-core==1.29.0
google-api-python-client==2.7.0
google-auth==1.30.1
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-speech==2.4.0
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
grpcio==1.38.0
httplib2==0.19.1
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.4
libcst==0.3.19
moviepy==1.0.3
mypy-extensions==0.4.3
numpy==1.20.3
packaging==20.9
Pillow==8.2.0
proglog==0.1.9
proto-plus==1.18.1
protobuf==3.17.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-decouple==3.4
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
tqdm==4.61.0
typing-extensions==3.10.0.0
typing-inspect==0.6.0
uritemplate==3.0.1
urllib3==1.26.5
